requirements.txt
flask
flask-cors
flask-jwt-extended
psycopg2-binary
pandas
scikit-learn
joblib
python-dotenv